Melhorias nas telas de cadastro e adicao de mensagens.
