+++ .settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,10 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <faceted-project>
-  <runtime name="Apache Tomcat v8.0"/>
   <fixed facet="java"/>
   <fixed facet="jst.web"/>
   <fixed facet="wst.jsdt.web"/>
-  <installed facet="java" version="1.8"/>
   <installed facet="jst.web" version="3.1"/>
   <installed facet="wst.jsdt.web" version="1.0"/>
   <installed facet="jboss.m2" version="1.0"/>
@@ -12,4 +10,5 @@
   <installed facet="jst.jsf" version="2.2"/>
   <installed facet="jst.jaxrs" version="2.0"/>
   <installed facet="jst.cdi" version="1.1"/>
+  <installed facet="java" version="1.8"/>
 </faceted-project>
